7/2 - Clean up code and mocks, disable Get Started on create, split code into smaller components
