+++ react-client/src/components/ViewJobPosting.tsx
